Move dynamic sampling functionality to its own library crate
