+++ relay-server/Cargo.toml
@@ -44,8 +44,6 @@ listenfd = "0.3.3"
 minidump = { git = "https://github.com/luser/rust-minidump", rev = "20070798b5e9216474013341ed80c5272558d9ee", optional = true }
 native-tls = { version = "0.2.4", optional = true }
 parking_lot = "0.10.0"
-rand_pcg="0.1.2"
-rand= "0.6.5"
 rdkafka = { version = "0.24", optional = true }
 rdkafka-sys = { version = "2.1.0", optional = true }
 regex = "1.3.9"
@@ -57,6 +55,7 @@ relay-general = { path = "../relay-general" }
 relay-log = { path = "../relay-log" }
 relay-quotas = { path = "../relay-quotas" }
 relay-redis = { path = "../relay-redis" }
+relay-dyn-sampling = { path = "../realy-dyn-sampling" }
 reqwest = { version = "0.10.8", features = ["gzip", "stream"] }
 rmp-serde = "0.14.3"
 serde = { version = "1.0.114", features = ["derive"] }
@@ -69,7 +68,6 @@ tokio = "0.2" # in sync with reqwest
 tokio-timer = "0.2.13"
 url = { version = "2.1.1", features = ["serde"] }
 uuid = { version = "0.8.1", features = ["v5"] }
-unicase="2.6.0"
 
 [target."cfg(not(windows))".dependencies]
 libc = "0.2.71"
